backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wavegauge-backend"
version = "0.2.3"
description = "Backend service for WaveGauge"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "numpy",
    "pandas",
    "asteval",
    "python-multipart",
    "websockets",
    "wavekit>=0.2",
]

[project.optional-dependencies]
gui = [
    "pywebview",
    "PyGObject; sys_platform == 'linux'",
]
dev = [
    "mypy>=1.8",
    "ruff>=0.6",
    "pyinstaller",
]

[tool.setuptools]
py-modules = ["app", "engine"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]

[tool.mypy]
python_version = "3.9"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_return_any = true
ignore_missing_imports = true
show_error_codes = true
exclude = "(^|/)build/"
